lorem ipsum!
i've finally populated all the pages!!!! granted, 99% of them are lipsum. with the use of the <base> tag and a lot of superfluous self-congratulation, i've gotten all the links to work with nothing but copying and pasting the same template - which imports a jquery script to put header and footer code where they should be - and then changing the base amount of ../s. i'm currently feeling really proud of myself, but i'm sure in a couple of years i'm going to regret doing all of this manually, esp with the individual news articles...